Changed Q from int to float

+++ module 0b find_tmin.py
@@ -50,7 +50,7 @@ elif check.lower() == 'n':
     C_l_vals.append(C_l)
     
 # flow rate
-Q = int(input('Enter flow rate: ')) * (10**27 / 3600) # mL/h to pm^3/s
+Q = float(input('Enter flow rate: ')) * (10**27 / 3600) # mL/h to pm^3/s
 
 # Trackmate files
 # nonspec_track_data = input('For flow rate = %d, enter name of \"Track statistics\" file(s) (from Trackmate) for non-specific ligand: ' % Q)
